+++ benchmarks/timeseries/resampling.py
@@ -7,15 +7,13 @@ from collections.abc import Sequence
 from datetime import datetime, timedelta, timezone
 from timeit import timeit
 
-from frequenz.quantities import Quantity
-
 from frequenz.sdk.timeseries import ResamplerConfig
 from frequenz.sdk.timeseries._resampling._base_types import SourceProperties
 from frequenz.sdk.timeseries._resampling._resampler import _ResamplingHelper
 
 
 def nop(  # pylint: disable=unused-argument
-    samples: Sequence[tuple[datetime, Quantity]],
+    samples: Sequence[tuple[datetime, float]],
     resampler_config: ResamplerConfig,
     source_properties: SourceProperties,
 ) -> float:
@@ -43,7 +41,7 @@ def _benchmark_resampling_helper(resamples: int, samples: int) -> None:
         for _n_resample in range(resamples):
             for _n_sample in range(samples):
                 now = now + timedelta(seconds=1 / samples)
-                helper.add_sample((now, Quantity(0.0)))
+                helper.add_sample((now, 0.0))
             helper.resample(now)
 
     print(timeit(_do_work, number=5))